Minor tweaks in ppeparation for release.
	modified:   ../../Makefile
	modified:   ../c99/Makefile
	modified:   ../c99/cpp.go
	modified:   Makefile
	modified:   all_test.go
	modified:   ../../sqlite2go.go

+++ internal/c99/cpp.go
@@ -362,7 +362,7 @@ func (c *cpp) parse(src ...Source) (tokenReader, error) {
 					encBuf = append(encBuf, encBuf1[:n]...)
 				}
 				id := dict.ID(encBuf)
-				if id > math.MaxUint32 {
+				if int64(id) > math.MaxUint32 {
 					panic("internal error 4")
 				}
